% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{bbKMPlot}
\alias{bbKMPlot}
\title{R coalesce Function}
\usage{
bbKMPlot(daten, gruppe = NULL, titel, filename, survtime = NULL,
  survtimetext = NULL, legendeout = NULL, file.out = TRUE,
  legfontsize = NULL, subtext = "", xmax = 61.25,
  xlab = "Monate (Anzeige bis max. 5 Jahre)", cex.lab = 1, cex.axis = 1,
  watermark = TRUE, ylab = "", logrank = FALSE)
}
\arguments{
\item{list}{of vectors of identical length}
}
\value{
one vector with the first non-NA value
}
\description{
accepts a list of vectors of identical length and returns one vector with the first non-NA value
}
\examples{
coalesce(c(NA,NA,53))

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

