% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{bbRelSurvCohortPlot}
\alias{bbRelSurvCohortPlot}
\title{R coalesce Function}
\usage{
bbRelSurvCohortPlot(x, ylab = NULL, xlab = NULL, titel = NULL,
  jahrstart = NULL, jahrende = NULL, jahrintervall = 1,
  method = "edererII", lang = "de")
}
\arguments{
\item{list}{of vectors of identical length}
}
\value{
one vector with the first non-NA value
}
\description{
accepts a list of vectors of identical length and returns one vector with the first non-NA value
}
\examples{
coalesce(c(NA,NA,53))

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

