% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{bbTheme}
\alias{bbTheme}
\title{R fuction to set a unified layout for base plots}
\usage{
bbTheme()
}
\description{
accepts a list of vectors of identical length and returns one vector with the first non-NA value
}
\examples{
bbTheme()

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

