% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{drop.levels}
\alias{drop.levels}
\title{Drop unused factor levels from all factors in a data.frame}
\usage{
drop.levels(dat)
}
\arguments{
\item{dat}{: data.frame}
}
\value{
dataframe
}
\description{
Drop unused factor levels from all factors in a data.frame
}
\examples{
df<-drop.levels(df)

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

