% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getAgeGroup}
\alias{getAgeGroup}
\title{Returns an age group for a given age in years}
\usage{
getAgeGroup(age, br)
}
\arguments{
\item{age}{the age in years}

\item{br}{the intervall in years to build groups on}
}
\value{
Vector of single character string containing a string like '20-30'
}
\description{
Returns an age group for a given age in years
}
\examples{
getAgeGroup(53,10)

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

