% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getBBColors}
\alias{getBBColors}
\title{R Function für generating a custom colorpallette of length x}
\usage{
getBBColors(x)
}
\arguments{
\item{x}{:numeric vector - size of colorpalette}
}
\value{
one character vector of length x containig the custom colors
}
\description{
R Function für generating a custom colorpallette of length x
}
\examples{
cols<-getBBColors(7)

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

