% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%w/o\%}
\alias{\%w/o\%}
\title{R not in / without operator}
\usage{
x \%w/o\% y
}
\arguments{
\item{x}{vector: the values not to be}

\item{y}{vector: the values to be searched against}
}
\value{

}
\description{
%w/o% is a more intuitive interface as a binary operator, which returns a logical vector indicating if there is no match for its left operand
}
\examples{
c(1:6,7:2) \%w/o\% c(3,7,12)

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

