% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{kmsurv}
\alias{kmsurv}
\title{Function to calculate the probability of survival for a point in time}
\usage{
kmsurv(S, totaltimes)
}
\value{
one vector with the first non-NA value
}
\description{
accepts a list of vectors of identical length and returns one vector with the first non-NA value
}
\examples{
kmsurv(S,M)

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

