% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{new.df}
\alias{new.df}
\title{Function to create an empty named data.frame}
\usage{
new.df(header)
}
\arguments{
\item{header}{vector of strings: The names for the columns of the data.frame}
}
\value{
A data.frame
}
\description{
Function to create an empty named data.frame
}
\examples{
ndf<-new.df(c("ID","NAME","AGE"))

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

