% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{p.value.survdiff}
\alias{p.value.survdiff}
\title{Function to exctract p-Value from surfdiff object}
\usage{
p.value.survdiff(daten, group = NULL, time = "time", status = "status")
}
\arguments{
\item{daten}{Data.Frame with survival datacontaining min. two columns with a) survival time b) survival status}

\item{group}{columnsname to be groubed by}

\item{time}{name of column containing the intervall data. default="time"}

\item{status}{name of colmn containig status data. default="status"}
}
\value{
a numeric vector representing the p-value
}
\description{
accepts a list of vectors of identical length and returns one vector with the first non-NA value
}
\examples{
\dontrun{
pvalue<-p.value.survdiff(daten=df,group="Treatment",time="duration",status="survivalstatus")
}

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
