% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{roughUICC}
\alias{roughUICC}
\title{Returns the rough UICC Stage from a detailed UICC Stage}
\usage{
roughUICC(uiccfull)
}
\arguments{
\item{uiccfull}{the detailed uicc stage}
}
\value{
Vector of single character string containing the rough UICC Stage
}
\description{
Returns the rough UICC Stage from a detailed UICC Stage
}
\examples{
roughUICC("IIIa")

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

