% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{bbBoxPlot}
\alias{bbBoxPlot}
\title{bbBoxPlot - streamlined BoxPlot}
\usage{
bbBoxPlot(daten, gruppe = NULL, filename = NULL, ylab, titel = NULL)
}
\arguments{
\item{daten}{the vector of numeric data}

\item{gruppe}{if given the factors to be grouped by}

\item{filename}{if given output will be delivered to file}

\item{ylab}{label of y axis}

\item{titel}{title of plot}
}
\value{
a base boxplot
}
\description{
R Function to streamline the generation of (grouped) boxplots
}
\examples{
bp<-bbBoxPlot(iris$Petal.Width,gruppe=iris$Species,ylab="Petal Width")

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
