% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{bbRelSurvCohortPlot}
\alias{bbRelSurvCohortPlot}
\title{Function to generate a Plot of 1-5 yera relative survival for different cohortes}
\usage{
bbRelSurvCohortPlot(x, ylab = NULL, xlab = NULL, titel = NULL,
  jahrstart = NULL, jahrende = NULL, jahrintervall = 1,
  method = "edererII", lang = "de")
}
\arguments{
\item{x}{survival data}

\item{ylab}{label of y axis}

\item{xlab}{label of x axis}

\item{jahrstart}{year bto start the intervall}

\item{jahrende}{year to en the intervall}

\item{jahrintervall}{stepsize of the intervall to be displayed default is one year}

\item{method}{to calculate relative survival default ='edererII'}

\item{title}{title of plot}
}
\value{
a plot
}
\description{
Function to generate a Plot of 1-5 yera relative survival for different cohortes
}
\examples{
relsurvcoplot<-bbRelSurvCohortPlot(x,ylab=NULL,xlab=NULL,titel=NULL,jahrstart=NULL,jahrende=NULL,jahrintervall=1,method='edererII', lang='de')

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

