% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{bbRelSurvPlot}
\alias{bbRelSurvPlot}
\title{R Function to plot the relative 1-5 years survival by a given group}
\usage{
bbRelSurvPlot(x, gruppe = NULL, ylab = NULL, xlab = NULL, titel = NULL,
  jahrstart = NULL, jahrende = NULL)
}
\arguments{
\item{x}{dataframe with survival data}

\item{gruppe}{the factor to be grouped by}

\item{ylab}{label of y axis}

\item{xlab}{label of x axis}

\item{jahrstart}{year to start the intervall}

\item{jahrende}{year of end of intervall}

\item{title}{title of plot}
}
\value{
a base plot
}
\description{
accepts a list of vectors of identical length and returns one vector with the first non-NA value
}
\examples{
relsurvplot<-bbRelSurvPlot(x,gruppe=NULL,ylab=NULL,xlab=NULL,titel=NULL,jahrstart=NULL,jahrende=NULL)

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

