% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{kmsurv}
\alias{kmsurv}
\title{Function to calculate the probability of survival for a point in time}
\usage{
kmsurv(S, totaltimes)
}
\value{
one vector with the first non-NA value
}
\description{
Function to calculate the probability of survival for a point in time
}
\examples{
x<-kmsurv(S,totaltimes)

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
